internship_pred/requirements.txt
Flask>=2.0.2
numpy>=1.21.4
pandas>=1.3.4
scikit-learn>=1.0.1
mysql-connector-python>=8.0.27
